Lazy initialization of coarse time goroutine
Initializing the coarseTime goroutine in the init() call had a pretty
serious side effect. It was impossible to override the value of
`time.Local` without causing a race condition, because its value is used
on `NewTicker` calls.

This commit changes the previous behavior by only initializing the
coarseTime goroutine when the first server is created. Allowing programs
to possibly override time.Local before mgo is initialized.