Count in-flight connection attempts towards pool limit
The connection pool limit can be exceeded at the moment if
server.Connect takes a long time, since in-flight connection attempts do
not count towards the pool limit calculation.

This caused a real problem for us because when our Mongo server started
slowing down due to load, connection attempts started taking a long time
(particuarly if a TLS handshake needs to be done) and so our processes
started opening many more actual TCP sockets than the configured
PoolLimit (thousands more). This of course made the Mongo server
situation worse in a cascading failure that should have been prevented
by the PoolLimit.

This patch fixes the issue by adding an "inFlightSocketDials" counter
which is incremented/decremented while connection attempts are in
progress, and counting that towards the "is the pool limit reached"
calculation.