Cleanup txn-queues even for only Asserted documents.
This has a threshold of how many transactions are ready to be removed
from the transaction queue. Once we reach that threshold, we go ahead
and update the document, even though it wouldn't require updates
otherwise.
In performance testing, it can be seen that having a document whose
queue grows too long can significantly impact how long it takes to apply
a new transaction. (We have to load the information about all the other
transactions that asserted against this document, and check if any of
them need to be applied.)

As an example, looping over 600 transactions to be applied, and timing
how long it takes to apply them with various AssertionCleanupLength
gives:

  1 1.18s
  2 1.17s
  5 1.28s
 10 1.39s
 20 1.81s
 50 2.94s
100 4.81s
  0 22.99s # unlimited

So keeping the queues small significantly impacts transaction speed.